WP CLI command added:
wp generate_welsh_providers

Run it from the project root directory (the one with the 'web' folder in it)

It successfully does the following for every provider in the system:

Sets the PLL language to English, in preparation for the rest:

Creates a Welsh version (copies the English title), copies over all the meta fields, and assigns the appropriate Welsh versions of the taxonomy terms i.e. sectors, local authorites, and provider statusses.